Adding functionality: saving incomes with quasi float type of amount

+++ FinanceManager.h
@@ -34,7 +34,7 @@ public:
     void addIncome();
     void displayIncomes(); //development only
     Income provideIncomeDetails();
-    bool amountCorrect(int amount);
+    bool amountCorrect(float amount);
 };
 
 #endif
